+++ tests/isolate/package_root_test.dart
@@ -17,7 +17,7 @@ main([args, port]) async {
       packageRoot: Uri.parse(SPAWN_PACKAGE_ROOT));
   p.handler = (msg) {
     p.close();
-    if (msg != SPAWN_PACKAGE_ROOT) {
+    if (msg != null) {
       throw "Bad package root in child isolate: $msg";
     }
     print("SUCCESS");
@@ -28,5 +28,5 @@ main([args, port]) async {
 testPackageRoot(port) async {
   var packageRoot = await Isolate.packageRoot;
   print("Spawned isolate's package root: $packageRoot");
-  port.send(packageRoot.toString());
+  port.send(packageRoot);
 }